Move app.js to components, create color-from.js
