+++ src/components/ChatListItem.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { withStyles } from '@material-ui/core/styles';
 import ListItem from '@material-ui/core/ListItem';
 import ListItemText from '@material-ui/core/ListItemText';
-import Avatar from '@material-ui/core/Avatar';
+import Avatar from './Avatar';
 
 import titleInitials from '../utils/title-initials';
 
@@ -12,7 +12,7 @@ const styles = theme => ({
 
 const ChatListItem = ({ classes, title }) => (
   <ListItem button>
-    <Avatar>{titleInitials(title)}</Avatar>
+    <Avatar colorFrom={title}>{title}</Avatar>
     <ListItemText primary={title}/>
   </ListItem>
 );
