+++ src/components/ChatMessage.js
@@ -2,10 +2,9 @@ import React from 'react';
 import classnames from 'classnames';
 import { withStyles } from '@material-ui/core';
 import Typography from '@material-ui/core/Typography';
-import Avatar from '@material-ui/core/Avatar';
+import Avatar from './Avatar';
 import Paper from '@material-ui/core/Paper';
 
-import titleInitials from '../utils/title-initials';
 
 const styles = theme => ({
   messageWrapper: {
@@ -33,8 +32,8 @@ const ChatMessage = ({ classes, sender, content }) => {
   const isMessageFromMe = sender === 'me';
 
   const userAvatar = (
-    <Avatar>
-      {titleInitials(sender)}
+    <Avatar colorFrom={sender}>
+      {sender}
     </Avatar>
   );
